analyzer: silence macOS double-precision warning on ISO_NAME rangeMax
UINT64_MAX (2^64-1) is not exactly representable as a double, so assigning
it to the double rangeMax member in the ISO_NAME_FIELD macro triggered
~20 clang -Wimplicit-const-int-float-conversion warnings on macOS (one per
PGN that contains an ISO Name field).

The value is harmless: ISO_NAME prints via fieldPrintName (not the numeric
path that reads rangeMax), and analyzer-explain.c already special-cases this
field to print the exact UINT64_MAX integer rather than the double. The
stored double is only a non-NaN marker.

Make the rounding explicit with (double) UINT64_MAX to silence the warning.

+++ analyzer/pgn.h
@@ -498,7 +498,7 @@ typedef struct
   {.name = nam, .size = BITS(2), .resolution = 1, .fieldType = "UNSIGNED_INTEGER", .hasSign = false, .rangeMin = 0, .rangeMax = 3}
 
 #define ISO_NAME_FIELD(nam) \
-  {.name = nam, .size = BYTES(8), .resolution = 1, .fieldType = "ISO_NAME", .hasSign = false, .rangeMin = 0, .rangeMax = UINT64_MAX}
+  {.name = nam, .size = BYTES(8), .resolution = 1, .fieldType = "ISO_NAME", .hasSign = false, .rangeMin = 0, .rangeMax = (double) UINT64_MAX}
 
 #define INSTANCE_FIELD \
   {.name = "Instance", .size = BYTES(1), .resolution = 1, .description = NULL, .partOfPrimaryKey = true, .fieldType = "UINT8"}
